improve code (move calculateDaysRemaining to resources file and remove unneccessary fragement
